Integrated Stripe & Added Webhooks

+++ src/main/java/org/example/paymentservice/paymentgateway/IPaymentGateway.java
@@ -1,5 +1,5 @@
 package org.example.paymentservice.paymentgateway;
 
 public interface IPaymentGateway {
-    String getPaymentLink(Double amount, Long orderId, String name, String phoneNumber);
+    String getPaymentLink(Long amount, Long orderId, String name, String phoneNumber);
 }
